[xunit.console] Move ConsoleRunner creation to a separate function
On platforms where Mono is the dotnet SDK runtime (Linux s390x, currently),
work around a limitation of the Mono JIT by moving the instantiation of
ConsoleRunner to a separate function from the call to
SubscribeResolveForAssembly.

Unlike CoreCLR, when the Mono JIT is compiling Main and sees a call to new
ConsoleRunner, it needs to load the types of the fields of ConsoleRunner.  Some
of those fields come from assemblies that can only be found using the resolve
helper.  But the helper is not installed until the JIT has finished compiling
Main and starts executing it.  This is a fundamental limitation of the Mono JIT
and is unlikely to be fixed in the short term.

The workaround is to move the code that depends on the assembly resolve to a
separate non-inlinable function.  As a result, Mono with JIT and invoke Main
without trying to load and initialize ConsoleRunner until after the assembly resolve
event handler is installed.

Fixes https://github.com/dotnet/runtime/issues/60550